chapter_003/src/main/java/sort/SortUser.java
package sort;

import java.util.ArrayList;
import java.util.List;
import java.util.Set;
import java.util.TreeSet;

/**
 * Package for class .
 *
 * @author Maxim Ignashov (mailto:ignashov.m@icloud.com)
 * @version 1.0
 * @since 24.08.2018
 */
public class SortUser {
    public Set<User> sort(List<User> list) {
        Set<User> treeSet = new TreeSet<>();
        for (User usrlst: list) {
            treeSet.add(new User(usrlst.getAge(), usrlst.getName()));
        }
        return treeSet;
    }
}
